==> code/00_scraping_DOLh2a.ipynb <==
## packages and imports

import os
import requests
from bs4 import BeautifulSoup
print(os.getcwd())

## pathname parameter
## going two-levels up from the current directory which is nested within Dropbox
OUTPUT_DIR = "../../qss20_finalproj_rawdata/summerwork/raw/"
REPULL_RAW = True

url = 'https://www.dol.gov/agencies/eta/foreign-labor/performance'

response = requests.get(url)

soup = BeautifulSoup(response.text, 'html.parser')
excel_files = soup.select('a[href*=xlsx]')

year = 2021

# NOTE: following code partially sourced from https://stackoverflow.com/questions/56459371/download-xls-file-from-a-webpage-web-scraper
# loop through all excel files
for ex in excel_files:
    # select h2a and skip addendum
    if 'H-2A' in ex.text and 'Addendum' not in ex.text:

        if year >= 2014:
            link = 'https://www.dol.gov' + ex['href']

            filename = "H_2A_Disclosure_Data_FY"+str(year)+".xlsx"

            if os.path.isfile(OUTPUT_DIR + '/' + filename):
                print ('*** File already exists: %s ***' %filename)
                year-=1
                continue


            resp = requests.get(link)
            output = open(OUTPUT_DIR + '/' + filename, 'wb')
            output.write(resp.content)
            output.close()
            print ('Saved: %s' %filename)

            year-=1

        # when reached 2014 data
        else:
            break
