==> code/08_clean_TRLA_intake.Rmd <==
---
title: "trla_cleaning"
author: "Lizzie Shackney"
date: "8/13/2021"
output: html_document
---


# Packages / imports

```{r}
library(tidyverse)
library(stringr)
library(fastLink)
library(readr)
library(data.table)
library(readxl)
library(here)
library(lubridate)


## Path to data
RUN_FROM_CONSOLE = FALSE
if(RUN_FROM_CONSOLE){
  args <- commandArgs(TRUE)
  DATA_DIR = args[1]
} else{
  DATA_DIR = "C:/Users/Austin Paralegal/Dropbox/qss20_finalproj_rawdata/summerwork"
  #DATA_DIR = "~/Dropbox/qss20_finalproj_rawdata/summerwork"
}

```

Load data

```{r warning=FALSE}
## orig data 
trla_orig = read_excel(sprintf("%s/raw/Adverse_Parties_FW_TRLA.xlsx", DATA_DIR)) %>%
    mutate(`Intake Date` = ymd(`Intake Date`))

## load second set of data
trla_more = read_excel(sprintf("%s/raw/master consol - redacted 2.xls", DATA_DIR)) %>%
  select(-c("closed","date_close"))

```


Clean up and rename:
```{r}
#clean column names in original
clean_1 = gsub("\\s+", "_", tolower(trimws(colnames(trla_orig))))
clean_2 = gsub("\\#", "number", clean_1)
clean_3  = gsub("(?!\\_)[[:punct:]]", "", clean_2, perl = TRUE)
colnames(trla_orig) = clean_3

#Rename trla_more columns to match orig
trla_more <- trla_more %>%
  rename(case_number = fileno, legal_problem_code = problem, case_disposition = appsts, intake_date = date_intrv, derived_concat_opp = opponent)

```


Info:
```{r}
sprintf("Columns in original data pull: %s", paste(colnames(trla_orig), collapse = ";")) 

sprintf("Columns in updated data pull: %s", paste(colnames(trla_more), collapse = ";")) 

sprintf("There are %s unique case numbers, %s unique case id/lead case ids, and %s rows in original dataset",
        length(unique(trla_orig$case_number)),
        length(unique(trla_orig$caseidlead_case_number)),
        nrow(trla_orig))

sprintf("There are %s unique case numbers and %s rows in second-pull dataset",
        length(unique(trla_more$case_number)),
        nrow(trla_more))

```


Consolidate AP columns:
```{r}
## focus on consolidating opponent columns

#If multiple are present you’d want to separate out.
#So, if ind AP is missing, use Lead AP. You have to decide if a “count” is tied to an individual client or to a group case.
#If Lead AP is missing, that is fine, use individual case AP.
#If both are present, you could separate out into two cases for the first occurrence? Or for each depending on how you’ve decided to count.

trla_orig = trla_orig %>%
      mutate(derived_opponent_consolidated = case_when(!is.na(adverse_party_organization) ~ adverse_party_organization,
                                               is.na(adverse_party_organization) & !is.na(lead_case_ap_organization) ~ lead_case_ap_organization,
                                               is.na(adverse_party_organization) & is.na(lead_case_ap_organization) &
                                              !is.na(adverse_party_name) ~ adverse_party_name,
                                              is.na(adverse_party_organization) & is.na(lead_case_ap_organization) &
                                              is.na(adverse_party_name) & !is.na(lead_case_ap_name) ~ lead_case_ap_name,
                                              TRUE ~ NA_character_),
             derived_opponent_source = case_when(!is.na(adverse_party_organization) ~ "AP org",
                                               is.na(adverse_party_organization) & !is.na(lead_case_ap_organization) ~ "Lead AP org",
                                               is.na(adverse_party_organization) & is.na(lead_case_ap_organization) &
                                              !is.na(adverse_party_name) ~ "AP name",
                                              is.na(adverse_party_organization) & is.na(lead_case_ap_organization) &
                                              is.na(adverse_party_name) & !is.na(lead_case_ap_name) ~ "Lead AP name",
                                              TRUE ~ "Missing all"),
             derived_opponent_state = (case_when(!is.na(ap_state) ~ ap_state,
                                                 is.na(ap_state) ~ lead_case_ap_name,
                                                 TRUE ~ NA_character_
                                                 )),
             derived_opponent_city = (case_when(!is.na(ap_city) ~ ap_city,
                                                 is.na(ap_city) ~ lead_case_ap_city,
                                                 TRUE ~ NA_character_
                                                 )),
             
             derived_is_notemp = case_when(grepl("Social Security Administration|Department of Labor|Workforce|Department|Dept|U.S. Immigration and Customs Enforcement|United States Customs and Border Protection|United States Environmental Protection Agency|United States Immigration and Naturalization Service|United States of America|Texas Attorney General|Texas Employment Commission|US Citizenship And Immigration Services|	
Texas Unemployment Commission|workforce", derived_opponent_consolidated) ~ TRUE,
                                      TRUE ~  FALSE)) %>%
  select(-c("date_closed", "date_of_rejection","date_opened"))
```


Working on the merge...

How many case numbers are shared between trla_orig with "Missing all" and trla_more?
How many will get filled in with the 'trla_more'?

Both a rowbind and a join.
1: If case number matches across sets and opp is missing, sub 'more'
2. If case number matches across sets and opp is present but different, add new row with 'more' info
3. If case number matches across sets and opp is present and same, do nothing
4. If case number is unique to trla_more, add new row


```{r}


#Create flags for 'in more' & 'missing AP' in either or both
#Filter out earlier data

trla_orig2 <- trla_orig %>%
  group_by(case_number) %>%
  mutate(in_more = ifelse(case_number %in% trla_more$case_number, TRUE, FALSE)) %>%
  ungroup() %>%
  mutate(derived_missing_AP = ifelse(is.na(derived_opponent_consolidated), TRUE, FALSE)) %>%
  filter(intake_date >= ymd("2014-01-01"))


sprintf("Of the %s in TRLA first pull missing an AP, %s are in trla more with non missing opponent",
        nrow(trla_orig2 %>% filter(derived_missing_AP)),
        nrow(trla_orig2 %>% filter(derived_missing_AP & case_number %in% trla_more$case_number[!is.na(trla_more$derived_concat_opp)])))

#See a count of various combinations of missing AP in orig (1) and more (2)
# missing_combos <- trla_orig2 %>%
#   distinct(case_number,.keep_all = TRUE) %>%
#   bind_rows(trla_more %>% rename(derived_opponent_consolidated = derived_concat_opp), .id = "id") %>%
#   select("case_number","id","missing_AP") %>%
#   pivot_wider(id_cols = "case_number", names_from = "id", values_from = "missing_AP") %>%
#   group_by(`1`,`2`) %>%
#   summarise(count = n()) %>%
#   ungroup()
# 
#   
# print(sprintf("There are %s case numbers across both datasets with no AP in either set. There are %s case numbers in trla_orig with a match in trla_more. There are %s case numbers in trla_more but not in trla_orig. There are %s cases with missing AP in trla_orig before merging.",
#         missing_combos %>%
#         filter((`1` & `2`) | (`1` & is.na(`2`)) | (`2` & is.na(`1`))) %>%
#           summarise(sum(count)),
#         missing_combos %>%
#         filter((`1` &!`2`)) %>%
#           summarise(sum(count)),
#         missing_combos %>%
#           filter(is.na(`1`)) %>%
#           summarise(sum(count)),
#         trla_orig2 %>%
#           filter(missing_AP) %>%
#           summarise(length(missing_AP))
#         ))

```


## More cleaning of trla_more before rowbind to try to convert to longform with oponne


```{r}
## try separating on the \r 
trla_more_longform = trla_more %>% mutate(derived_opponent_consolidated = strsplit(derived_concat_opp, "\r"), # split opponent at R
                                     derived_opponent_state = strsplit(opstate, "\\,"), # split oppstate at ,
                                     derived_nopp = lengths(derived_opponent_consolidated), # number of opponents
                                     derived_nstate = lengths(derived_opponent_state), # number of states
                                     derived_is_difflengths = case_when(derived_nopp != derived_nstate ~ TRUE, # true if n opp != n states
                                                                     TRUE ~ FALSE)) %>%
                filter(!derived_is_difflengths) %>% # filter out for now to get unnest -> longform to work but shown below
                unnest(c(derived_opponent_consolidated, derived_opponent_state)) 


#View(trla_more %>% filter(!case_number %in% trla_more_sep$case_number)) # lose 84 that don't have that delimiter
```


```{r}
#Follow proposed logic to join the datasets

#filter dataset with only unique trla_more rows
trla_more_notinorig <- trla_more_longform %>%
  anti_join(trla_orig2, by = c("case_number")) %>%
  filter(intake_date >= ymd("2014-01-01")) %>%
  mutate(derived_dfsource = "trla_more",
         derived_opponent_source = "trla_more")


# first, using case numbers that overlap, merge
trla_merge_overlap <- trla_orig2 %>%
  
  #If case number matches across sets and opp is missing, sub 'more'
  #Join with matching case numbers in trla_more to fill in missing opponent info
  left_join(trla_more_longform %>%
          dplyr::select(-derived_nopp, -derived_nstate, 
                        -derived_is_difflengths,
                        -derived_concat_opp,
                        -opstate), by = "case_number", suffix = c("_orig", "_more")) %>%
  
  mutate(
    
    # create new opponent column that uses orig if present; more if orig missing 
    derived_opponent_consolidated = case_when(!is.na(derived_opponent_consolidated_orig) ~ derived_opponent_consolidated_orig,
                                                   is.na(derived_opponent_consolidated_orig) ~ derived_opponent_consolidated_more,
                                              TRUE ~ NA_character_),
    
    # similar with state (more doesnt have city so cant use similar logic
    derived_opponent_state = case_when(!is.na(derived_opponent_state_orig) ~ derived_opponent_state_orig,
                                        is.na(derived_opponent_state_orig) ~ derived_opponent_state_more,
                                              TRUE ~ NA_character_),
         
    #note source of AP
    derived_dfsource = case_when(is.na(derived_opponent_consolidated) ~ "missing in both",
                                        is.na(derived_opponent_consolidated_orig) ~ "trla_more",
                                        TRUE ~ "trla_orig"),
    
    derived_opponent_source = case_when((derived_opponent_source == "Missing all" & !is.na(derived_opponent_consolidated)) ~ "trla_more",
                                        !is.na(derived_opponent_source) ~ derived_opponent_source,
                                        (is.na(derived_opponent_source) & derived_dfsource =="trla_more") ~ "trla_more",
                                        TRUE ~ "Missing all")) 

## then, rowbind that with ones not present in original
trla_merge_all = trla_merge_overlap %>%
          bind_rows(trla_more_notinorig, .id = "source_rows") %>%
          mutate(derived_dfsource = ifelse(source_rows == 2, "trla_more_uniquecase", derived_dfsource)) 



```

```{r}
## within a case_number, create flags for at least one present
trla_merge_all = trla_merge_all %>%
          group_by(case_number) %>%
          mutate(derived_allopp_miss = ifelse(sum(is.na(derived_opponent_consolidated)) == n(), TRUE, FALSE),
                 derived_allstate_miss = ifelse(sum(is.na(derived_opponent_state)) == n(), TRUE, FALSE)) %>%
          ungroup() 

sprintf("Out of %s rows, there are %s missing opponent", nrow(trla_merge_all), sum(is.na(trla_merge_all$derived_opponent_consolidated)))
sprintf("Out of %s unique case numbers, there are %s missing all opponents", length(unique(trla_merge_all$case_number)),
                        nrow(trla_merge_all %>% dplyr::select(case_number, derived_allopp_miss) %>% distinct() %>% filter(derived_allopp_miss)))
```

**rj note**: assume we should write longform data with all rows of the same case number retained. but not sure if we should group by case number and fill some of the missing fields with other values for that group (or group by both case_number and opponent)?

Write to csv/dropbox
```{r}

table(trla_merge_all$derived_opponent_state[!trla_merge_all$derived_allopp_miss]) # states look relatively clean for ones not missing

## write to dropbox ones with non-missing opponents
trla_merge_nonmiss = trla_merge_all %>% filter(!derived_allopp_miss & !derived_allstate_miss) %>%
        mutate(derived_legalproblemcode = ifelse(is.na(legal_problem_code_orig), legal_problem_code_more, legal_problem_code_orig),
               derived_intakedate = case_when(is.na(intake_date_orig) ~ as.Date(intake_date_more),
                                              !is.na(intake_date_orig) ~ as.Date(intake_date_orig),
                                              TRUE ~ NA_Date_),
               derived_casedispo = case_when(!is.na(case_disposition_orig) ~ case_disposition_orig,
                                             case_disposition_more == "CLO" ~ "Closed",
                                             case_disposition_more == "PEN" ~ "Pending",
                                             case_disposition_more == "ACC" ~ "Open",
                                             TRUE ~ "Unknown dispo"
                                             )) %>%
        filter(!derived_is_notemp) %>% # filter out the institutional opponents discussed earlier
        dplyr::select(caseidlead_case_number, case_number, derived_intakedate, derived_casedispo, derived_opponent_consolidated,
               derived_opponent_state, derived_opponent_city, derived_legalproblemcode) 

trla_merge_nonmiss <- trla_merge_nonmiss %>%
  distinct(case_number, derived_opponent_consolidated, derived_opponent_state, derived_intakedate, .keep_all = TRUE)


saveRDS(trla_merge_nonmiss, paste0(DATA_DIR, "intermediate/cleaned_TRLA_formerge.RDS"))
write.csv(trla_merge_nonmiss, paste0(DATA_DIR, "intermediate/cleaned_TRLA_formerge.csv"))


saveRDS(trla_merge_nonmiss, "cleaned_TRLA_formerge.RDS")
write.csv(trla_merge_nonmiss, "cleaned_TRLA_formerge.csv")

```
